+++ launch_chffrplus.sh
@@ -91,10 +91,21 @@ function launch {
     "$BASEDIR/installer/updater/updater" "file://$BASEDIR/installer/updater/update.json"
   fi
 
-  # Remove old NEOS update file
-  # TODO: move this code to the updater
+  # Remove old NEOS update files
   if [ -d /data/neoupdate ]; then
     rm -rf /data/neoupdate
+
+  # One-time fix for a subset of OP3T with gyro orientation offsets.
+  # Remove and regenerate qcom sensor registry. Only done on OP3T mainboards.
+  # Performed exactly once. The old registry is preserved just-in-case, and
+  # doubles as a flag denoting we've already done the reset.
+  # TODO: we should really grow per-platform detect and setup routines
+  if ! $(grep -q "letv" /proc/cmdline) && [ ! -f "/persist/comma/op3t-sns-reg-backup" ]; then
+    echo "Performing OP3T sensor registry reset"
+    mv /persist/sensors/sns.reg /persist/comma/op3t-sns-reg-backup &&
+      rm -f /persist/sensors/sensors_settings /persist/sensors/error_log /persist/sensors/gyro_sensitity_cal &&
+      echo "restart" > /sys/kernel/debug/msm_subsys/slpi &&
+      sleep 5  # Give Android sensor subsystem a moment to recover
   fi
 
   # handle pythonpath